[Bronze V] Title: A+B - 9, Time: 124 ms, Memory: 21400 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15740. A＋B － 9/A＋B － 9.kt
@@ -1,1 +1,1 @@
-fun main() = readln().split(" ").map { it.toInt() }.let { println(it.sum()) }
+fun main() = readln().split(" ").map { it.toLong() }.let { println(it.sum()) }
